code/device-healthiness-monitoring-service/ad-pipeline/pyproject.toml
[tool.poetry]
name = "ad-pipeline"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
confluent-kafka = "^2.3.0"
pydantic = "^2.6.3"
pydantic-settings = "^2.2.1"
loguru = "^0.7.2"
pendulum = "^3.0.0"
pysad = "^0.2.0"
httpx = "^0.27.0"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
